split Recommend and RecommendContext
