+++ packages/recommend-react/src/index.ts
@@ -7,5 +7,4 @@ export * from './useRecommendations';
 export * from './useRelatedProducts';
 export * from './useTrendingFacets';
 export * from './useTrendingItems';
-export { Recommend } from './Recommend';
-export type { RecommendProps } from './Recommend';
+export * from './Recommend';
